backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0
greenlet==3.0.1
# ── 文件解析 ──────────────────────────────────────────────────
pdfplumber>=0.10.0        # PDF 文字提取
easyocr>=1.7.0            # 图片本地 OCR（中英文，首次运行自动下载模型，无 API Key 时兜底）
Pillow>=10.0.0            # 图像处理
numpy>=1.24.0             # easyocr 依赖
# ── AI 大模型识别（推荐，精度更高） ────────────────────────────
openai>=1.0.0             # DashScope OpenAI 兼容接口
pymupdf>=1.24.0           # 扫描 PDF 转图片（PyMuPDF）
python-dotenv>=1.0.0      # 从 .env 文件加载环境变量
